backend/requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
boto3==1.21.26
botocore==1.24.26
certifi==2021.10.8
diff-match-patch==20200713
et-xmlfile==1.1.0
jmespath==1.0.0
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.9
Pillow==9.1.0
s3transfer==0.5.2
six==1.16.0
tablib==3.1.0
tzdata==2022.1
urllib3==1.26.9
wincertstore==0.2
xlrd==2.0.1
xlwt==1.3.0
cement==2.8.2
colorama==0.4.3
Django==4.1.1
django-cors-headers==3.13.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
future==0.16.0
pathspec==0.9.0
PyJWT==2.4.0
pytz==2022.2.1
semantic-version==2.8.5
sqlparse==0.4.2
termcolor==1.1.0
tzdata==2022.2
wcwidth==0.1.9
mysqlclient==2.1.0